refactor: moved build_model() from train.py to framework.py
also some code cleanup and simplification
